Adjust lookup of startup files to respect ZDOTDIR
This change allows the startup file lookup to succeed in case a user has
specified an alternative base directory for their zsh startup files.

In case the `ZDOTDIR` environment variable is set, lookups for zsh
config files use that as a base instead of home[1]. This change replicates
that behaviour.

[1] https://wiki.archlinux.org/index.php/Zsh#Startup.2FShutdown_files